src/requirements.txt
requests==2.25.1
pandas==1.2.4
pyyaml==5.4.1
